add aug

+++ data/dataset.py
@@ -78,7 +78,7 @@ class Video(list):
             ret, frame = cap.read()
             if ret:
                 # Our operations on the frame come here
-                gray = cv2.resize(cv2.cvtColor(frame, cv2.COLOR_BGR2GRAY), (120, 120)).reshape(-1, 1, 120, 120).astype(np.float64)
+                gray = cv2.resize(cv2.cvtColor(frame, cv2.COLOR_BGR2GRAY), (120, 120)).reshape(-1, 1, 120, 120)
                 tmp.append(gray)
                 # Display the resulting frame
                 #imshow(gray, cmap='gray')
